there is a collective communication bug concerning integers
workaround: use unsigned long.

the bug for integers still needs to be fixed.



git-svn-id: https://scm.projects.hlrs.de/svn/ls1/MarDyn/branches/density-profiles@2821 a63bd714-7e14-4b5e-94e7-302c8c8ff188

+++ src/parallel/DomainDecomposition.h
@@ -180,7 +180,7 @@ public:
 	}
 
 	float collCommGetFloat() {
-		return _collComm.getInt();
+		return _collComm.getFloat();
 	}
 
 	double collCommGetDouble() {
